backend/requirements.txt
#Son todas la librerias necesarias para que nuestro proyecto funcione
Flask==2.1.1
python-dotenv
flask-restful
flask_sqlalchemy
flask-jwt-extended
Flask-Mail